+++ src/controllers/enrollmentReports/index.ts
@@ -1,5 +1,5 @@
 export { parseUploadedTemplate, parseEnrollmentReportRow } from './parse';
 export * from './removeExisting';
 export * from './checkErrors';
-export * from './map/mapUtils';
+export * from './map/mapRows/utils';
 export * from './formatUploadPreview';
