+++ src/controllers/enrollmentReports/map/mapRows/utils/fundingTime.ts
@@ -1,6 +1,6 @@
-import { FundingSource } from '../../../../../client/src/shared/models';
-import { FUNDING_SOURCE_TIMES } from '../../../../../client/src/shared/constants';
-import { normalizeString } from '../../../../utils/normalizeString';
+import { FundingSource } from '../../../../../../client/src/shared/models';
+import { FUNDING_SOURCE_TIMES } from '../../../../../../client/src/shared/constants';
+import { normalizeString } from '../../../../../utils/normalizeString';
 
 const FT_REGEX = /fulltime/g;
 const PT_REGEX = /parttime/g;
