+++ src/controllers/enrollmentReports/map/mapRows/utils/index.ts
@@ -5,4 +5,3 @@ export * from './family';
 export * from './funding';
 export * from './fundingTime';
 export * from './incomeDetermination';
-export * from './site';
