+++ src/index.ts
@@ -1,4 +1,5 @@
 import express, { json } from 'express';
+
 import path from 'path';
 import httpProxy from 'http-proxy';
 import { createConnection, getConnectionOptions } from 'typeorm';
